improve: custom cli extenesion
