+++ packages/cli/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@axiomkit/cli",
-  "version": "0.0.7",
+  "version": "0.0.8",
   "type": "module",
   "main": "./dist/index.js",
   "module": "./dist/index.js",
